fix: allow float duration in VideoMetadata to support Instagram Reels
- Updated Pydantic model VideoMetadata to accept float durations
- Added Instagram Reel test case to regression suite

+++ backend/main.py
@@ -96,7 +96,7 @@ class PlaylistEntry(BaseModel):
 
 class VideoMetadata(BaseModel):
     title: str
-    duration: Optional[int] = None
+    duration: Optional[float] = None
     thumbnail: Optional[str] = None
     formats: List[FormatInfo]
     url: str
